Support int96 type predicate push down
